requirements.txt
fitz
numpy
sentence-transformers
transformers
torch
chromadb
requests
frontend
tools
google-generativeai